backend/requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
passlib[bcrypt]
python-jose
stripe
python-dotenv
